refactor: migrate TalkBox types from seller to common and update related imports
