+++ src/pages/seller/talkBox/questionsList/PendingQuestionsTab.tsx
@@ -6,7 +6,7 @@ import { SubCategoryChip, InfiniteQuestionList } from '@/components';
 import {
   QuestionDTO,
   SingleQuestionAnswerDTO,
-} from '@/types/seller/TalkBox.types';
+} from '@/types/common/TalkBox.types';
 
 // store
 import {
