+++ src/services/talkBox/mutation/useGetUserTalkboxList.tsx
@@ -1,4 +1,4 @@
-import { getUserQuesionList } from '@/api/talkBox/getUserQuesionList.api';
+import { getUserQuestionList } from '@/api/talkBox/getUserQuestionList.api';
 import { QUERY_KEYS } from '@/constants/api';
 import { useInfiniteQuery } from '@tanstack/react-query';
 
@@ -6,7 +6,7 @@ export const useGetUserTalkboxList = () => {
   const query = useInfiniteQuery({
     queryKey: [QUERY_KEYS.USER_ITEM_TALKBOX_LIST],
     queryFn: ({ pageParam = 1 }: { pageParam: number }) =>
-      getUserQuesionList({ page: pageParam, size: 15 }),
+      getUserQuestionList({ page: pageParam, size: 15 }),
     getNextPageParam: (lastPage, allPages) => {
       // lastPage: 마지막으로 불러온 페이지의 데이터
       // allPages: 지금까지 불러온 모든 페이지의 배열
