--- src/types/common/TalkboxType.types.ts
@@ -1,10 +1,0 @@
-export interface TalkboxType {
-  itemId: number;
-  itemTitle: string;
-  itemMainPic: string;
-  sellerNickname: string;
-  sellerProfilePic: string | null;
-  lastChatContent: string;
-  lastChatTime: string;
-  uncheckedCnt: number;
-}
